[Backout] - CL33543358 [FYI] jordi.rovira Original CL Desc ----------------------------------------------------------------- [mutable] Restore the layout ID redesign. Fix some related bugs.
- Mesh cloning with flags wasn't replicating the MeshIdPrefix.
- Mesh reformat was sometimes duplicating the system buffers.
- Fix bugs with the wrapping project handling of layout block ids

#jira UE-209986
[REVIEW] [at]pere.rifa
#rnx
#rb pere.rifa

[CL 33570911 by bob tellez in ue5-main branch]

+++ Engine/Plugins/Experimental/Mutable/Source/MutableTools/Private/MuT/ASTOpMeshExtractLayoutBlocks.h
@@ -26,8 +26,8 @@ namespace mu
 		/** Layout to use to select the blocks. */
 		uint16 Layout = 0;
 
-		/** Block Ids to include in the resulting mesh. */
-		TArray<uint64> Blocks;
+		/** Blocks to include in the resulting mesh. */
+		TArray<uint32> Blocks;
 
 	public:
